# api/specifications/vitanexus_api_spec.yaml
openapi: 3.0.3
info:
  title: VitaNexus API
  description: |
    REST API for VitaNexus Health Assurance Cooperative.

    Core API for member management, health scoring, financial operations,
    and third-party integrations (wearables, EHR, claims processors).

    **Security**: All endpoints require OAuth 2.0 authentication with JWT tokens.
    PHI data is encrypted in transit (TLS 1.3) and at rest (AES-256).

    **Rate Limiting**: 1000 requests/minute per API key for standard tier.

    **HIPAA Compliance**: All PHI access is logged in audit trail.

  version: 1.0.0
  contact:
    name: VitaNexus API Support
    email: api-support@vitanexus.com
  license:
    name: Proprietary
servers:
  - url: https://api.vitanexus.com/v1
    description: Production server
  - url: https://api-staging.vitanexus.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development
security:
  - OAuth2:
      - read:members
      - write:members
      - read:health_data
      - write:health_data
      - read:financial
      - write:financial
...
paths:
...
  /members:
    post:
      tags:
        - Members
      summary: Enroll new member
      description: Create new member enrollment with demographic and health information
      operationId: createMember
      security:
        - OAuth2:
            - write:members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberEnrollmentRequest'
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - Members
      summary: List members
      description: Retrieve paginated list of members with optional filters
      operationId: listMembers
      security:
        - OAuth2:
            - read:members
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, terminated, pending]
        - name: risk_category
          in: query
          schema:
            type: string
            enum: [low, moderate, high, critical]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [enrollment_date, risk_score, last_name]
            default: enrollment_date
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemberSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
  /members/{member_id}:
    get:
      tags:
        - Members
      summary: Get member details
      operationId: getMember
      security:
        - OAuth2:
            - read:members
      parameters:
        - name: member_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Members
      summary: Update member information
      operationId: updateMember
      security:
        - OAuth2:
            - write:members
      parameters:
        - name: member_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdateRequest'
      responses:
        '200':
          description: Member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.vitanexus.com/oauth/authorize
          tokenUrl: https://auth.vitanexus.com/oauth/token
          scopes:
            read:members: Read member information
            write:members: Create and update members
            read:health_data: Read health data including PHI
            write:health_data: Write health data
            read:financial: Read financial data
            write:financial: Submit claims and process payments
  schemas:
    MemberEnrollmentRequest:
      type: object
      required:
        - first_name
        - last_name
        - date_of_birth
        - email
        - enrollment_date
      properties:
        external_member_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F, O]
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        enrollment_date:
          type: string
          format: date
        monthly_premium:
          type: number
          format: decimal
        initial_health_data:
          $ref: '#/components/schemas/HealthProfileUpdate'
    MemberResponse:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        external_member_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        age:
          type: integer
        gender:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        enrollment_date:
          type: string
          format: date
        status:
          type: string
        current_risk_score:
          type: number
          format: decimal
        current_risk_category:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MemberSummary:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        full_name:
          type: string
        age:
          type: integer
        risk_score:
          type: number
        risk_category:
          type: string
        enrollment_date:
          type: string
          format: date
        status:
          type: string
    MemberUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        status:
          type: string
          enum: [active, suspended, terminated]
    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
          minLength: 2
          maxLength: 2
        zip_code:
          type: string
        country:
          type: string
          default: USA
...
    HealthProfileUpdate:
      type: object
      properties:
        biometrics:
          type: object
        lab_results:
          type: object
        behaviors:
          type: object
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_records:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'